move a few things around as per Tom's suggestions
