+++ holopy/core/io/__init__.py
@@ -15,4 +15,4 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with HoloPy.  If not, see <http://www.gnu.org/licenses/>.
-from .io import load, load_image, save, save_image, get_example_data, get_example_data_path, load_average, bg_correct
+from .io import load, load_image, save, save_image, get_example_data, get_example_data_path, load_average
