+++ holopy/fitting/fit_series.py
@@ -31,7 +31,7 @@ import types
 import numpy as np
 from holopy.core.utils import mkdir_p
 from holopy.core.process import normalize, subimage
-from holopy.core import Image
+from holopy.core.metadata import Image
 from holopy.core.io import load, save
 from holopy.fitting import fit
